+++ rskj-core/src/test/java/co/rsk/rpc/Web3RskImplTest.java
@@ -84,6 +84,7 @@ public class Web3RskImplTest {
                 pm,
                 em,
                 tpm,
+                null,
                 Web3Mocks.getMockChannelManager(),
                 Web3Mocks.getMockRepository(),
                 null,
