+++ rskj-core/src/test/java/org/ethereum/rpc/Web3ImplScoringTest.java
@@ -366,6 +366,7 @@ public class Web3ImplScoringTest {
                 pm,
                 em,
                 tpm,
+                null,
                 Web3Mocks.getMockChannelManager(),
                 rsk.getRepository(),
                 peerScoringManager,
