+++ rskj-core/src/test/java/org/ethereum/rpc/Web3ImplTest.java
@@ -263,6 +263,7 @@ public class Web3ImplTest {
                 personalModule,
                 null,
                 txPoolModule,
+                null,
                 Web3Mocks.getMockChannelManager(),
                 Web3Mocks.getMockRepository(),
                 null,
@@ -956,6 +957,7 @@ public class Web3ImplTest {
                 personalModule,
                 null,
                 null,
+                null,
                 Web3Mocks.getMockChannelManager(),
                 Web3Mocks.getMockRepository(),
                 null,
@@ -1271,6 +1273,7 @@ public class Web3ImplTest {
                 personalModule,
                 ethModule,
                 txPoolModule,
+                null,
                 channelManager,
                 Web3Mocks.getMockRepository(),
                 null,
@@ -1321,6 +1324,7 @@ public class Web3ImplTest {
                 personalModule,
                 ethModule,
                 txPoolModule,
+                null,
                 channelManager,
                 Web3Mocks.getMockRepository(),
                 null,
@@ -1357,6 +1361,7 @@ public class Web3ImplTest {
                 personalModule,
                 ethModule,
                 txPoolModule,
+                null,
                 Web3Mocks.getMockChannelManager(),
                 Web3Mocks.getMockRepository(),
                 null,
@@ -1407,6 +1412,7 @@ public class Web3ImplTest {
                 new PersonalModuleWalletDisabled(),
                 ethModule,
                 txPoolModule,
+                null,
                 Web3Mocks.getMockChannelManager(),
                 Web3Mocks.getMockRepository(),
                 null,
